+++ src/api/configuration.cpp
@@ -5,8 +5,8 @@
 // The full license is in the file LICENSE, distributed with this software.
 
 #include "mamba/api/configuration.hpp"
+#include "mamba/api/create.hpp"
 #include "mamba/api/info.hpp"
-#include "mamba/api/install.hpp"
 
 #include "mamba/core/environment.hpp"
 #include "mamba/core/fetch.hpp"
